finish the documentation with swagger setup
