+++ src/lib/index.ts
@@ -1,0 +1,2 @@
+export { default as getApiDocs } from "./swagger/swagger";
+export { default as ReactSwagger } from "./swagger/SwaggerUI";
