+++ src/lib/swagger/SwaggerUI.tsx
@@ -8,7 +8,6 @@ type Props = {
 };
 
 function ReactSwagger({ spec }: Props) {
-  console.log(spec);
   return <SwaggerUI spec={spec} />;
 }
