+++ src/lib/swagger/swagger.ts
@@ -1,0 +1,25 @@
+import { createSwaggerSpec } from "next-swagger-doc";
+
+export default async function getApiDocs() {
+  const spec = createSwaggerSpec({
+    apiFolder: "src/app/api", // define api folder under app folder
+    definition: {
+      openapi: "3.0.0",
+      info: {
+        title: "Next Swagger API Example",
+        version: "1.0",
+      },
+      components: {
+        securitySchemes: {
+          BearerAuth: {
+            type: "http",
+            scheme: "bearer",
+            bearerFormat: "JWT",
+          },
+        },
+      },
+      security: [],
+    },
+  });
+  return spec;
+}
